requirements.txt
delta-rest-client
pandas
python-dotenv
requests
